requirements.txt
# General
pandas
openpyxl
PyMuPDF==1.25.3
google-generativeai
python-dotenv
requests
numpy
httpx

# Django Framework
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-celery-results
celery
redis

# Database
psycopg2-binary

# Authentication & Security
bcrypt
passlib[bcrypt]
sendgrid
certifi
python-jose[cryptography]
